requirements.txt
# Project-Cortex v2.0 - RPi 5 (ARM64) / Python 3.11
# Fixed: Jan 22, 2026

# ===== Core Framework =====
python-dotenv>=1.0.1        # Env management

# ===== Computer Vision =====
# CRITICAL: Python 3.13 mandates Numpy 2.1+. Legacy Numpy 1.x will fail.
numpy>=2.1.0                
opencv-python-headless>=4.10.0.84  # Headless for Lite OS; supports Numpy 2.x

# NOTE: Install picamera2 via apt: 'sudo apt install python3-picamera2'
# picamera2>=0.3.0

# ===== AI/ML Models =====
# Torch 2.6+ is required for Python 3.13 ARM64 support
ultralytics>=8.4.0          # YOLOv11/v10
torch>=2.6.0                # PyTorch for ARM64 (CPU)
torchvision>=0.21.0         # Matching vision library
pillow>=11.1.0              # Image processing
ncnn>=1.0.20240410           # NCNN inference
onnxruntime>=1.17.0          # ONNX inference
onnx>=1.15.0                 # ONNX export support


# ===== Cloud AI APIs =====
#zai-sdk>=0.2.1              # FIXED: Latest official Z.ai SDK (was zhipuai)
google-genai>=1.60.0        # Unified SDK for Gemini 2.0 Flash
google-generativeai>=0.8.6  # Legacy Gemini support
openai>=1.60.1              # OpenAI Client
gradio-client>=1.7.0        # Hugging Face Interface
requests>=2.32.3            # HTTP Client

# ===== Local AI Models =====
openai-whisper>=20240930    # STT (requires system ffmpeg)
kokoro-onnx>=0.4.9          # Local TTS (ONNX backend)
misaki[en]>=0.5.0           # G2P for Kokoro
# NOTE: Do NOT install 'phonemizer' — it conflicts with 'phonemizer-fork' needed by kokoro-onnx.
# Install phonemizer-fork separately: pip install phonemizer-fork>=3.3.2
espeakng-loader>=0.1.3      

# ===== OCR (Hailo NPU + PaddleOCR detection) =====
paddlepaddle>=2.5.0         # PaddlePaddle CPU-only (detection backbone)
paddleocr>=2.7.0            # PaddleOCR text detection (CPU, recognition on Hailo NPU)

# ===== Audio Processing =====
pygame>=2.6.1               # Audio I/O
sounddevice>=0.5.1          # Recording
scipy>=1.15.0               # Signal processing (Numpy 2.x ready)
pydub>=0.25.1               # Audio manipulation
silero-vad>=5.1.2           # Voice Activity Detection

# ===== Web & Dashboard =====
fastapi>=0.115.6            
uvicorn[standard]>=0.34.0   
websockets>=14.1            
pyzmq>=26.0.0               # ZeroMQ for Video Stream
imagezmq>=1.2.0             # ZMQ helper            

# ===== Cloud Storage =====
supabase>=2.3.4              # PostgreSQL + Storage for Layer 4 Memory

# ===== Utilities =====
pyyaml>=6.0.2               
psutil>=6.1.1               
